isolate generic function into a utils.js file
